libft version02

+++ ft_strlen.c
@@ -1,11 +1,11 @@
-#include <stddef.h>
+#include "libft.h"
 
 size_t ft_strlen(char *c)
 {
-    int i;
+    size_t i;
 
     i = 0;
-    while(c[i])
+    while (c[i])
         i++;
     return (i);
 }
